+++ src/main/java/io/nagurea/smsupsdk/common/response/APIOutputStreamResponse.java
@@ -5,13 +5,11 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.io.OutputStream;
-
 @Getter
 @AllArgsConstructor
 @ToString
 @EqualsAndHashCode
-public class APIOutputStreamResponse<T extends OutputStream>{
+public class APIOutputStreamResponse<T extends Document<?>>{
 
     private final String uid;
     private final Integer statusCode;