+++ src/main/java/io/nagurea/smsupsdk/common/response/Document.java
@@ -1,0 +1,17 @@
+package io.nagurea.smsupsdk.common.response;
+
+import lombok.Getter;
+import lombok.ToString;
+
+import java.io.OutputStream;
+
+@ToString
+@Getter
+public class Document<T extends OutputStream> {
+
+    protected final T documentOutputStream;
+
+    public Document(T documentOutputStream) {
+        this.documentOutputStream = documentOutputStream;
+    }
+}