+++ src/main/java/io/nagurea/smsupsdk/invoices/get/download/DownloadInvoiceService.java
@@ -2,6 +2,7 @@ package io.nagurea.smsupsdk.invoices.get.download;
 
 
 import io.nagurea.smsupsdk.common.http.get.GETSMSUpService;
+import io.nagurea.smsupsdk.common.response.PDFDocument;
 import io.nagurea.smsupsdk.helper.URLHelper;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 
@@ -26,8 +27,8 @@ public class DownloadInvoiceService extends GETSMSUpService {
      * @throws IOException when something got wrong during effective query to SMSUp
      */
     public DownloadInvoiceResponse downloadInvoice(String token, String id) throws IOException {
-        final ImmutablePair<Integer, OutputStream> response = getPDF(buildUrlById(id), token);
-        final OutputStream pdf = response.getRight();
+        final ImmutablePair<Integer, PDFDocument> response = getPDF(buildUrlById(id), token);
+        final PDFDocument pdf = response.getRight();
         return DownloadInvoiceResponse.builder()
                 .uid(UUID.randomUUID().toString())
                 .statusCode(response.getLeft())